requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2020.4.5.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.5
cycler==0.11.0
Django==4.1.1
docker-pycreds==0.4.0
easydict==1.8
fonttools==4.37.3
gitdb==4.0.9
GitPython==3.1.28
idna==3.4
image==1.5.24
kiwisolver==1.4.4
matplotlib==3.6.0
numpy==1.23.3
packaging==21.3
pathtools==0.1.2
Pillow==7.1.2
promise==2.3
protobuf==4.21.7
psutil==5.9.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==6.0
requests==2.28.1
scipy==1.4.1
sentry-sdk==1.9.10
setproctitle==1.3.2
shortuuid==1.0.9
six==1.14.0
smmap==5.0.0
sqlparse==0.4.3
timm==0.6.7
torch==1.9.0
torchvision==0.10.0
tqdm==4.64.1
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==16.0.0
wandb==0.13.4
